# New division implementation based on uint64 words