@holiman 2020-03-17
Sweet -- once this is in, I believe that this lib is faster than big.Int in every single op -- and in most places by a very wide margin!