@chfast 2020-03-17
Added documentation. Fixup commits to be squashed before merge.